requirement.txt
Django
djangorestframework
pytest



pytest-django
mypy
django-stubs
djangorestframework-stubs
black
isort